# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Trivela API
  version: 0.1.0
  description: |
    REST API for the Trivela campaign and rewards platform. Handles campaign metadata,
    health checks, and Soroban RPC configuration for the frontend.

    See the [API Migration Guide](../docs/API_MIGRATION.md) for v0 → v1 breaking changes,
    endpoint changelog, deprecation policy, and migration examples with curl snippets.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.trivela.example.com
    description: Production server
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/campaigns/{id}/stats:
    get:
      summary: Get campaign analytics and stats
      tags:
        - Campaigns
      operationId: getCampaignStats
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
        - name: range
          in: query
          description: Time range (7d, 30d, all)
          schema:
            type: string
            enum: [7d, 30d, all]
            default: 7d
        - name: from
          in: query
          description: Start date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: End date (ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Campaign statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignStats'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/index/campaigns/{id}/participants:
    get:
      summary: Cursor-paginated participant list from indexed balances
      tags: [IndexRead]
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: cursor
          schema: { type: string }
          description: Opaque pagination cursor from previous response
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Paginated participant list with freshness indicator
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        address: { type: string }
                        balance: { type: string }
                  cursor: { type: string, nullable: true }
                  has_more: { type: boolean }
                  as_of_ledger: { type: integer, nullable: true }
        '404':
          description: Campaign not found
  /api/v1/index/campaigns/{id}/events:
    get:
      summary: Cursor-paginated credit/claim event log for a campaign
      tags: [IndexRead]
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string, enum: [credit, claim] }
          description: Filter by event type
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Paginated event list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        user: { type: string }
                        amount: { type: string }
                        ledger: { type: integer }
                        tx_hash: { type: string }
                        type: { type: string, enum: [credit, claim] }
                  cursor: { type: string, nullable: true }
                  has_more: { type: boolean }
                  as_of_ledger: { type: integer, nullable: true }
        '404':
          description: Campaign not found
  /api/v1/index/addresses/{address}/history:
    get:
      summary: Full event history for a single Stellar address
      tags: [IndexRead]
      security: []
      parameters:
        - in: path
          name: address
          required: true
          schema: { type: string }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Address event history and current balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  address: { type: string }
                  balance: { type: string }
                  data: { type: array }
                  cursor: { type: string, nullable: true }
                  has_more: { type: boolean }
                  as_of_ledger: { type: integer, nullable: true }
  /api/v1/index/campaigns/{id}/stats:
    get:
      summary: Rollup stats for a campaign (ETag-cached)
      tags: [IndexRead]
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Campaign stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaignId: { type: string }
                  as_of_ledger: { type: integer, nullable: true }
                  summary:
                    type: object
                    properties:
                      totalParticipants: { type: integer }
                      totalCredited: { type: string }
                      totalClaimed: { type: string }
                      claimRate: { type: number }
        '304':
          description: Not Modified (ETag match)
        '404':
          description: Campaign not found

  # ── #556: Sponsored accounts ───────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for write operations
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  schemas:
...
    Error:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
...
    CampaignStats:
      type: object
      required:
        - campaignId
        - onChainSynced
        - range
        - summary
        - registrationsByDay
        - pointsByDay
      properties:
        campaignId:
          type: string
          description: Campaign ID
        onChainSynced:
          type: boolean
          description: Whether on-chain data has been synced
        range:
          type: object
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
        summary:
          type: object
          required:
            - totalParticipants
            - totalPoints
            - claimRate
            - timeRemainingMs
          properties:
            totalParticipants:
              type: integer
              description: Total registered participants
            totalPoints:
              type: integer
              description: Total points awarded
            claimRate:
              type: number
              description: Percentage of points claimed
            timeRemainingMs:
              type: integer
              description: Milliseconds until campaign ends
        registrationsByDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              count:
                type: integer
          description: Daily registration counts
        pointsByDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              credited:
                type: integer
              claimed:
                type: integer
          description: Daily credited and claimed points